chore(sql-editor): update data table lite
