--- frontend/src/views/sql-editor/EditorCommon/ResultView/DataTable/DataTable/index.ts
@@ -1,3 +1,0 @@
-import DataTable from "./DataTable.vue";
-
-export default DataTable;
