--- frontend/src/views/sql-editor/EditorCommon/ResultView/DataTable/DataTableLite/index.ts
@@ -1,3 +1,0 @@
-import DataTableLite from "./DataTableLite.vue";
-
-export default DataTableLite;
