+++ frontend/src/views/sql-editor/EditorCommon/ResultView/DataTable/index.ts
@@ -1,4 +1,4 @@
-import DataTable from "./DataTable";
-import DataTableLite from "./DataTableLite";
+import DataTable from "./DataTable.vue";
+import DataTableLite from "./DataTableLite.vue";
 
 export { DataTable, DataTableLite };
